Initial work on FFmpeg Heroku Buildpack
The idea is to use Heroku Docker build image as an environment to build binaries. Rake is used to build them. AWS S3 SDK for Ruby is used to upload them.

When buildpack is used, it downloads those binaries from S3 and extracts them. Code for this has been copied and modified from JVM Heroku Buildpack. It is certain that this works is not complete since code hasn’t been tested and PATH environment variable needs to be modified to include FFmpeg binaries.